Housekeeping: remove unmaintained repos
I'm the owner of the denoserverless org and I've been doing some housekeeping in my repos. I recently renamed it to denofn, since I want to use the org to maintain [denoFn (now denofn-selfhosted)](https://github.com/jeroenptrs/denoFn). That has made me come to the conclusion that there's a ton of clutter on there that I added in this doc and I want to prevent devs that are new to Deno from using unmaintained software.

- jwt: djwt is maintained and works perfectly fine
- https-libs: all those dependencies can most likely be resolved via pika.dev/skypack

I already archived the repos but plan on removing them for the reasons listed above once they're out of this document.