Add with-env (#302)
Simple command line utility for executing commands with one or more .env files. Written with Deno and installed using the `deno install` command - though binaries are also distributed via GitHub.

This is different to existing entries as it is intended to be used as a stand-alone command line utility and not imported as a library.